+++ src/view/list-item.js
@@ -1,0 +1,3 @@
+export const createListItem = () => {
+  return '<li class="trip-events__item"></li>';
+};
